requirements.txt
aiohttp
aiosqlite
discord.py
requests
python-dotenv